+++ stb-remote/ULR1Handler.h
@@ -11,7 +11,7 @@ class ULR1Handler : public IRRemoteHandler
         ULR1Handler();
         virtual ~ULR1Handler();
         
-        virtual bool IdentifyRCUCommand(uint16_t command) override;
+        virtual bool IdentifyRCUCommand(uint32_t command) override;
         virtual bool HandleRCUCommand(String command, IRsend irsend) override;
     private:
         static std::map<uint32_t, String> m_hexToButtonMap;
